Extract specification management into seperate type
